requirements.txt
fastapi>=0.104,<1.0
uvicorn>=0.24,<1.0
pydantic>=2.5,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
tiktoken>=0.7,<1.0
matplotlib>=3.7,<4.0
